requirements.txt
Flask==2.3.3
pymongo==4.5.0
PyPDF2==3.0.1
nltk==3.8.1
scikit-learn==1.3.0
pandas==2.0.3
requests==2.31.0
flask-cors==4.0.0
Werkzeug==2.3.7
python-dotenv==1.0.0
